+++ CHANGELOG.md
@@ -4,6 +4,9 @@
     * Added support for stretching time on the timing dialog
     * The object context menu now can show a dilaog to change animation timing on the object
     * Ability to add fonts from Google Fonts
+* I/O:
+    * Fixed position of tspan on SVG output
+    * Shapes not showing on SVG output
 * Scripting:
     * Ability to stretch time for documents and individual objects
 * Bug Fixes: